move `ElGamalError` to the `errors` module
