+++ zk-token-sdk/src/errors.rs
@@ -1,7 +1,6 @@
 //! Errors related to proving and verifying proofs.
 #[cfg(not(target_os = "solana"))]
 use crate::{
-    encryption::elgamal::ElGamalError,
     range_proof::errors::{RangeProofGenerationError, RangeProofVerificationError},
     sigma_proofs::errors::*,
 };
@@ -17,6 +16,24 @@ pub enum AuthenticatedEncryptionError {
     SeedLengthTooLong,
     #[error("failed to deserialize")]
     Deserialization,
+}
+
+#[derive(Error, Clone, Debug, Eq, PartialEq)]
+pub enum ElGamalError {
+    #[error("key derivation method not supported")]
+    DerivationMethodNotSupported,
+    #[error("seed length too short for derivation")]
+    SeedLengthTooShort,
+    #[error("seed length too long for derivation")]
+    SeedLengthTooLong,
+    #[error("failed to deserialize ciphertext")]
+    CiphertextDeserialization,
+    #[error("failed to deserialize public key")]
+    PubkeyDeserialization,
+    #[error("failed to deserialize keypair")]
+    KeypairDeserialization,
+    #[error("failed to deserialize secret key")]
+    SecretKeyDeserialization,
 }
 
 #[cfg(not(target_os = "solana"))]
